- Fixed bug in image distort for I16 and S32 images - Tweaks related to convolution kernels - Improved README.TXT a bit

+++ main/ip/src/boofcv/alg/distort/impl/ImplImageDistort_I16.java
@@ -39,7 +39,7 @@ public class ImplImageDistort_I16<T extends ImageInt16> extends ImageDistortBasi
 
 	@Override
 	protected void assign(int indexDst, float value) {
-		dstImg.data[indexDst] = (byte)value;
+		dstImg.data[indexDst] = (short)value;
 	}
 
 }
